refactor: extract project/member domain to repository/project.rs
